+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Migrate to `node@6`.
+- Update node server configs.
+
 ## [1.17.1] - 2020-10-06
 
 ### Fixed